[resto druid] Missed Flourish in refactor
