+++ src/Parser/Druid/Restoration/README.md
@@ -3,7 +3,7 @@
 | --- | --- | --- |
 | [Always Be Casting](https://github.com/buimichael/RestoDruidAnalyzer/blob/master/src/Main/Parser/Modules/Features/AlwaysBeCasting.js) | Tracks your casting and healing time to determine your non healing and downtime. May get slightly inaccurate with unaccounted for Haste buffs. | 0% |
 | [Ability Tracker](https://github.com/buimichael/RestoDruidAnalyzer/blob/master/src/Main/Parser/Modules/Core/AbilityTracker.js) | Tracks ability casts and healing done (damage tracking NYI). | 100% |
-| [Flourish Analyzer](https://github.com/buimichael/RestoDruidAnalyzer/blob/master/src/Main/Parser/Modules/Features/Flourish.js) | Tracks flourish usage | 100% |
+| [Flourish Analyzer](https://github.com/buimichael/RestoDruidAnalyzer/blob/master/src/Main/Parser/Modules/Talents/Flourish.js) | Tracks flourish usage | 100% |
 | [Tree of Life throughput calculcation](https://github.com/buimichael/RestoDruidAnalyzer/blob/master/src/Main/Parser/Modules/Features/TreeOfLife.js) | Calculates Tree of Life throughput | 80-90% |
 | [Lifebloom uptime](https://github.com/buimichael/RestoDruidAnalyzer/blob/master/src/Main/Parser/Modules/Features/Lifebloom.js) | Get uptime on lifebloom | 100% |
 | [Efflorescence uptime](https://github.com/buimichael/RestoDruidAnalyzer/blob/master/src/Main/Parser/Modules/Features/Efflorescence.js) | Get uptime on efflorescence | 90-100% |
